Remove util classes
