+++ src/main/java/qupath/bioimageio/spec/tensor/InputTensor.java
@@ -1,10 +1,9 @@
 package qupath.bioimageio.spec.tensor;
 
-import qupath.bioimageio.spec.Processing;
-
 import java.util.List;
 
-import static qupath.bioimageio.spec.Utils.toUnmodifiableList;
+import static qupath.bioimageio.spec.Model.toUnmodifiableList;
+
 
 /**
  * Model input, including shape, axes, datatype and preprocessing.
