+++ src/main/java/qupath/bioimageio/spec/tensor/Processing.java
@@ -1,4 +1,20 @@
-package qupath.bioimageio.spec;
+/*
+ * Copyright 2025 University of Edinburgh
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package qupath.bioimageio.spec.tensor;
 
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
@@ -7,13 +23,13 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import qupath.bioimageio.spec.axes.Axis;
+import qupath.bioimageio.spec.tensor.axes.Axis;
 
 import java.lang.reflect.Type;
 import java.util.Collections;
 import java.util.Map;
 
-import static qupath.bioimageio.spec.Utils.deserializeField;
+import static qupath.bioimageio.spec.Model.deserializeField;
 
 /**
  * Base class for pre- and post-processing operations.
@@ -211,7 +227,7 @@ public class Processing {
     }
 
     public static class ProcessingModeDeserializer implements JsonDeserializer<ProcessingMode> {
-        Logger logger = LoggerFactory.getLogger(Processing.class);
+        private static final Logger logger = LoggerFactory.getLogger(Processing.class);
         @Override
         public Processing.ProcessingMode deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
                 throws JsonParseException {
@@ -247,7 +263,7 @@ public class Processing {
 
             var obj = json.getAsJsonObject();
             var name = obj.has("name") ? obj.get("name").getAsString() : obj.get("id").getAsString();
-            JsonObject kwargs = deserializeField(context, obj, "kwargs", JsonObject.class, null);;
+            JsonObject kwargs = deserializeField(context, obj, "kwargs", JsonObject.class, null);
             switch (name) {
                 case "binarize":
                     var binarize = new Processing.Binarize();
