+++ src/main/java/qupath/bioimageio/spec/tensor/axes/SpaceAxes.java
@@ -1,7 +1,23 @@
-package qupath.bioimageio.spec.axes;
+/*
+ * Copyright 2025 University of Edinburgh
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 
-import qupath.bioimageio.spec.sizes.ReferencedSize;
-import qupath.bioimageio.spec.sizes.Size;
+package qupath.bioimageio.spec.tensor.axes;
+
+import qupath.bioimageio.spec.tensor.sizes.ReferencedSize;
+import qupath.bioimageio.spec.tensor.sizes.Size;
 import qupath.bioimageio.spec.tensor.BaseTensor;
 
 import java.util.List;
@@ -42,11 +58,25 @@ public class SpaceAxes {
         SpaceUnit(String s) {
             this.unit = s;
         }
+
+        /**
+         * Get the SpaceUnit corresponding to a String value.
+         * @param unit The unit as a string
+         * @return The corresponding unit, or null if not found.
+         */
+        public static SpaceUnit getUnit(String unit) {
+            for (var su: SpaceUnit.values()) {
+                if (su.unit.equalsIgnoreCase(unit)) {
+                    return su;
+                }
+            }
+            return null;
+        }
     }
 
     abstract static class SpaceAxisBase extends AxisBase implements ScaledAxis {
         private final SpaceUnit unit;
-        private double scale = 1.0;
+        private double scale;
 
         SpaceAxisBase(String id, String description, String unit, double scale) {
             this(id, description, SpaceUnit.valueOf(unit.toUpperCase()), scale);
@@ -114,7 +144,7 @@ public class SpaceAxes {
 
     static class SpaceOutputAxisWithHalo extends SpaceOutputAxis implements WithHalo {
         private ReferencedSize size;
-        private int halo = 0;
+        private int halo;
 
         SpaceOutputAxisWithHalo(String id, String description, String unit, double scale, Size size, int halo) {
             super(id, description, unit, scale, size);
