+++ src/main/java/qupath/bioimageio/spec/tensor/sizes/Size.java
@@ -1,4 +1,20 @@
-package qupath.bioimageio.spec.sizes;
+/*
+ * Copyright 2025 University of Edinburgh
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package qupath.bioimageio.spec.tensor.sizes;
 
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonElement;
@@ -10,7 +26,7 @@ import qupath.bioimageio.spec.tensor.BaseTensor;
 
 import java.util.List;
 
-import static qupath.bioimageio.spec.Utils.deserializeField;
+import static qupath.bioimageio.spec.Model.deserializeField;
 
 
 /**
@@ -42,7 +58,7 @@ public interface Size {
 
     void validate(List<? extends BaseTensor> tensors);
 
-    public static Size deserializeSize(JsonDeserializationContext context, JsonElement jsonElement, JsonElement scale) {
+    static Size deserializeSize(JsonDeserializationContext context, JsonElement jsonElement, JsonElement scale) {
         Logger logger = LoggerFactory.getLogger(Size.class);
         if (jsonElement == null || jsonElement.isJsonNull()) {
             return null;
